Revert "change to_int32 to tf.cast"
This reverts commit dc3eaa67d22e4a0fe6cf4f0f92a62c4ad0b02c25.

+++ models-r1.13.0/official/transformer/model/ffn_layer.py
@@ -62,8 +62,7 @@ class FeedFowardNetwork(tf.layers.Layer):
         # Flatten padding to [batch_size*length]
         pad_mask = tf.reshape(padding, [-1])
 
-        #nonpad_ids = tf.to_int32(tf.where(pad_mask < 1e-9))
-        nonpad_ids = tf.cast(tf.where(pad_mask < 1e-9), tf.int32)
+        nonpad_ids = tf.to_int32(tf.where(pad_mask < 1e-9))
 
         # Reshape x to [batch_size*length, hidden_size] to remove padding
         x = tf.reshape(x, [-1, self.hidden_size])
